parallaxDrawer: added scroll preservation to site submitter parallax drawer
